Simplify mapping and remove useless metadata
