+++ package.json
@@ -17,7 +17,7 @@
     "axios": "1.11.0",
     "bcryptjs": "3.0.2",
     "cors": "2.8.5",
-    "dbsder-api-types": "3.0.1",
+    "dbsder-api-types": "3.0.3",
     "dotenv": "17.2.1",
     "express": "5.1.0",
     "joi": "18.0.1",
