+++ src/backend/modules/document/service/documentService/fetchAllJurisdictions.spec.ts
@@ -11,7 +11,6 @@ describe('fetchAllJurisdictions', () => {
           decisionMetadata: documentModule.decisionMetadataGenerator.generate({
             appealNumber: '',
             additionalTermsToAnnotate: '',
-            boundDecisionDocumentNumbers: [],
             categoriesToOmit: [],
             chamberName: '',
             criminalCaseCode: '',
