--- src/backend/modules/document/service/documentService/fetchDoneDocumentWithoutLossNotIn.ts
@@ -1,12 +1,0 @@
-import { documentType } from '@src/core';
-import { buildDocumentRepository } from '../../repository';
-
-export { fetchDoneDocumentWithoutLossNotIn };
-
-async function fetchDoneDocumentWithoutLossNotIn(
-  documentIdsToExclude: documentType['_id'][],
-): Promise<documentType | undefined> {
-  const documentRepository = buildDocumentRepository();
-
-  return await documentRepository.findOneByStatusWithoutLossNotIn(['done', 'toBePublished'], documentIdsToExclude);
-}
