+++ src/backend/modules/document/service/documentService/fetchTreatedDocuments.ts
@@ -17,7 +17,6 @@ async function fetchTreatedDocuments(settings: settingsType) {
       'decisionMetadata',
       'creationDate',
       'documentNumber',
-      'loss',
       'publicationCategory',
       'reviewStatus',
       'source',
@@ -75,7 +74,6 @@ async function fetchTreatedDocuments(settings: settingsType) {
         creationDate: treatedDocument.creationDate,
         documentNumber: treatedDocument.documentNumber,
         jurisdiction: treatedDocument.decisionMetadata.jurisdiction,
-        loss: treatedDocument.loss,
         occultationBlock: treatedDocument.decisionMetadata.occultationBlock,
         publicationCategory: treatedDocument.publicationCategory,
         reviewStatus: treatedDocument.reviewStatus,
