+++ src/core/modules/document/generator/documentGenerator.ts
@@ -38,7 +38,6 @@ const decisionMetadataGenerator: generatorType<documentType['decisionMetadata']>
     additionalTermsParsingFailed,
     computedAdditionalTerms,
     appealNumber,
-    boundDecisionDocumentNumbers,
     categoriesToOmit,
     chamberName,
     civilCaseCode,
@@ -59,7 +58,6 @@ const decisionMetadataGenerator: generatorType<documentType['decisionMetadata']>
     computedAdditionalTerms: computedAdditionalTerms ?? undefined,
     additionalTermsParsingFailed: additionalTermsParsingFailed ?? undefined,
     appealNumber: appealNumber ?? '',
-    boundDecisionDocumentNumbers: boundDecisionDocumentNumbers ?? [],
     categoriesToOmit: categoriesToOmit ?? [],
     chamberName: chamberName ?? '',
     civilCaseCode: civilCaseCode ?? '',
@@ -86,7 +84,6 @@ const documentGenerator: generatorType<documentType> = {
     externalId,
     _id,
     importer,
-    loss,
     priority,
     publicationCategory,
     reviewStatus,
@@ -105,7 +102,6 @@ const documentGenerator: generatorType<documentType> = {
     externalId: externalId ?? `EXTERNAL_ID_${Math.random()}`,
     _id: _id ? new ObjectId(_id) : new ObjectId(),
     importer: importer ?? 'default',
-    loss: loss,
     priority: priority !== undefined ? priority : 0,
     publicationCategory: publicationCategory ? publicationCategory : [],
     reviewStatus: reviewStatus || { hasBeenAmended: false, viewerNames: [] },
