--- src/core/modules/document/lib/computeCaseNumber.ts
@@ -1,10 +1,0 @@
-import { documentType } from '../documentType';
-
-export { computeCaseNumber };
-
-function computeCaseNumber(document: Pick<documentType, 'decisionMetadata' | 'documentNumber'>): number {
-  if (document.decisionMetadata.boundDecisionDocumentNumbers.length > 0) {
-    return document.decisionMetadata.boundDecisionDocumentNumbers[0];
-  }
-  return document.documentNumber;
-}
