+++ src/core/modules/document/lib/index.ts
@@ -1,6 +1,5 @@
 import { buildDocument } from './buildDocument';
 import { comparator } from './comparator';
-import { computeCaseNumber } from './computeCaseNumber';
 import { countWords } from './countWords';
 import { getNextStatus } from './getNextStatus';
 import { getMinutesBeforeFreeingPendingDocuments } from './getMinutesBeforeFreeingPendingDocuments';
@@ -9,7 +8,6 @@ import { publicationHandler } from './publicationHandler';
 export {
   buildDocument,
   comparator,
-  computeCaseNumber,
   countWords,
   getNextStatus,
   getMinutesBeforeFreeingPendingDocuments,
