+++ src/core/modules/ressourceFilter/lib/filterTreatedDocuments.spec.ts
@@ -167,7 +167,6 @@ describe('filterTreatedDocuments', () => {
           chamberName: '',
           date: decisionDate,
           categoriesToOmit: [],
-          boundDecisionDocumentNumbers: [],
           NACCode: '',
           endCaseCode: '',
           appealNumber: '',
