--- src/courDeCassation/connector/mapper/extractors/extractReadableJurisdictionName.spec.ts
@@ -1,19 +1,0 @@
-import { extractReadableJurisdictionName } from './extractReadableJurisdictionName';
-
-describe('extractReadableJurisdictionName', () => {
-  it('should return "Cour de cassation" ', () => {
-    const jurisdictionName = 'Cour de cassation';
-
-    const readableJurisdictionName = extractReadableJurisdictionName(jurisdictionName);
-
-    expect(readableJurisdictionName).toBe('Cour de cassation');
-  });
-
-  it('should return "cour d\'appel de Rouen" ', () => {
-    const jurisdictionName = "cour d'appel de Rouen";
-
-    const readableJurisdictionName = extractReadableJurisdictionName(jurisdictionName);
-
-    expect(readableJurisdictionName).toBe("cour d'appel de Rouen");
-  });
-});
