+++ src/courDeCassation/connector/mapper/extractors/index.ts
@@ -1,13 +1,4 @@
 import { extractReadableChamberName } from './extractReadableChamberName';
-import { extractReadableJurisdictionName } from './extractReadableJurisdictionName';
-import {
-  extractAppealRegisterRoleGeneralNumber,
-  regexExtractAppealNumber,
-} from './extractAppealRegisterRoleGeneralNumber';
+import { extractNumeroPourvoi, regexExtractAppealNumber } from './extractNumeroPourvoi';
 
-export {
-  extractReadableChamberName,
-  extractReadableJurisdictionName,
-  extractAppealRegisterRoleGeneralNumber,
-  regexExtractAppealNumber,
-};
+export { extractReadableChamberName, extractNumeroPourvoi, regexExtractAppealNumber };
